implement THEME_MODE migration and validation; update AppConfig and ThemeContext for theme management
