+++ src/renderer/src/contexts/AppConfigContextDef.ts
@@ -1,5 +1,6 @@
 import { createContext } from 'react'
 import type { AppConfig } from '@shared/types/configTypes'
+import type { ThemeMode } from '@shared/types/themeTypes'
 
 export interface AppConfigContextType {
   config: AppConfig | null
@@ -7,6 +8,7 @@ export interface AppConfigContextType {
   errors: string[]
   refreshConfig: () => Promise<void>
   updateConfig: (updates: Partial<AppConfig>) => void
+  updateThemeMode: (themeMode: ThemeMode) => Promise<void>
 }
 
 export const AppConfigContext = createContext<AppConfigContextType>({
@@ -14,5 +16,6 @@ export const AppConfigContext = createContext<AppConfigContextType>({
   isLoading: true,
   errors: [],
   refreshConfig: async () => {},
-  updateConfig: () => {}
+  updateConfig: () => {},
+  updateThemeMode: async () => {}
 })
