+++ src/renderer/src/contexts/ThemeContextValue.ts
@@ -5,7 +5,6 @@ export interface ThemeContextValue {
   theme: Theme
   themeMode: ThemeMode
   setThemeMode: (mode: ThemeMode) => void
-  toggleTheme: () => void
 }
 
 export const ThemeContext = createContext<ThemeContextValue | undefined>(undefined)
