+++ src/shared/themes/index.ts
@@ -1,0 +1,11 @@
+// Export all themes
+export { lightTheme } from './lightTheme'
+export { darkTheme } from './darkTheme'
+export { osrsTheme } from './osrsTheme'
+
+// Export the themes object and default theme
+export { themes, defaultTheme } from './themes'
+
+// Export types
+export type { Theme } from '@shared/types/themeTypes'
+export type { ThemeMode } from '@shared/types/themeTypes'
