+++ src/shared/types/configTypes.ts
@@ -1,8 +1,11 @@
+import type { ThemeMode } from './themeTypes'
+
 export interface AppConfig {
   BASE_LOG_DIRECTORY: string
   DREAMBOT_VIP_FEATURES: boolean
   BASE_WEBHOOK_URL: string
   BOT_CONFIG: Record<string, BotConfig>
+  THEME_MODE: ThemeMode
 }
 
 export interface BotConfig {
